requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
edge-tts
google-transliteration-api
google-genai
pydantic
